Rename fetcher module to provider
