+++ src/lib.rs
@@ -1,4 +1,4 @@
 pub mod data;
-pub mod fetcher;
 pub mod helper;
 pub mod model;
+pub mod provider;
